Updated to work with Minetest 5.2
Made some minor updates to the orginal archived code so it will now work with minetest_game included with Minetest engine 5.2.

Changes
added farming as a dependancy
added snowblock
added iceblock
adjusted a few lines of code to allow comboblock to work with the current version of the stairs mod.